+++ odoo/local-src/specific_payment_mode/wizard/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import account_payment_mode_generator
+from . import mail_compose_message